fix: make currency fields uint64

+++ core/event.go
@@ -18,14 +18,14 @@ func (this SymbolEvent) IsOnline() bool  { return this.EventUnixTime < 0 }
 func (this SymbolEvent) UnixNano() int64 { return this.EventUnixTime & math.MaxInt64 }
 
 type BookEvent struct {
-	Bid, Ask      uint32
+	Bid, Ask      uint64
 	EventUnixTime int64
 }
 
 func (this BookEvent) UnixNano() int64 { return this.EventUnixTime }
 
 type CandleEvent struct {
-	Open, High, Low, Close uint32
+	Open, High, Low, Close uint64
 	EventUnixTime          int64
 }
 
@@ -38,8 +38,8 @@ func (this TradeEvent) UnixNano() int64 { return int64(this) & math.MaxInt64 }
 
 type ScheduleEvent struct {
 	Symbol                 uint64
-	Balance, Basis         uint32
-	MinOrder, MinIncrement uint32
+	Balance, Basis         uint64
+	MinOrder, MinIncrement uint64
 	EventUnixTime          int64
 }
